fix(*): fix components before sidebar
