+++ gemini-utils/ok-icon/index.tsx
@@ -1,0 +1,7 @@
+import OkIcon from './ok-icon';
+import OkColorIcon from './ok-icon-color';
+
+export {
+    OkIcon,
+    OkColorIcon,
+};
